saccosphere-project/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg
python-decouple
dj-database-url
psycopg2-binary
gunicorn
whitenoise
Pillow
redis
sentry-sdk
celery
